rename package to supabase_auth
