+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "gotrue"
+name = "supabase_auth"
 version = "2.4.1"
 description = "Python Client Library for Supabase Auth"
 authors = ["Joel Lee <joel@joellee.org>"]
@@ -36,7 +36,7 @@ python-semantic-release = "^9.1.1"
 pygithub = ">=1.57,<3.0"
 
 [tool.semantic_release]
-version_variables = ["gotrue/__init__.py:__version__"]
+version_variables = ["supabase_auth/__init__.py:__version__"]
 version_toml = ["pyproject.toml:tool.poetry.version"]
 major_on_zero = false
 commit_message = "chore(release): bump version to v{version}"
