+++ supabase_auth/constants.py
