+++ supabase_auth/errors.py
